Fix crash with user-defined routers
When adding a router to a given environment, the router object pointer
is passed to the CLIPS context via `ffi.new_handle`.

If the user does not retain the router object, it will be garbage
collected. This means that the handle created via `ffi.new_handle` is
freed leaving a dangling pointer.

When the router is then invoked it retrieves the dangling pointer from
the Environment context crashing.

This also caused some memory leak as the router was not properly
garbage collected once an Environment was de-allocated.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>