Fix constructs memory handling
Most of CLIPS constructs are kept within the engine's data structures
up until they are not deleted/undefined.

Once so, their pointers become dangling and any further API call
will lead to segfaults or memory corruptions.

Previously, this was a limitation which lead to unstable APIs.

The new implementation does not store the constructs pointers anymore
but keeps a reference to their identifiers. Any time the User
interacts with a construct, the internal pointer is retrieved from the
engine making sure the construct is still existing.

In case the construct is deleted/undefined but the Python object
retained, any interaction with it will lead to an exception
rather than a crash.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>