issue #48, functions: improve Python defined function error handling
When exceptions happen in a Python defined function, the execution of
the Rule is interrupted. The error and traceback are logged on
`stdwrn` routers and the error message is forwarded back to CLIPS.

A new set of API `error_state` and `clear_error_state()` are provided
to check if errors during rule execution happened.

Signed-off-by: Matteo Cafasso <noxdafox@gmail.com>